requirements.txt
pdf2image==1.17.0
plotly==5.24.1
python_pptx==1.0.2
streamlit==1.37.1
